% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ispv-package.R
\docType{package}
\name{ispv-package}
\alias{ispv}
\alias{ispv-package}
\title{ispv: Use Czech labour market survey data}
\description{
Retrieve and load data from the Czech Information System on Average Earnings (ISPV) at <https://www.ispv.cz>.
}
\author{
\strong{Maintainer}: Petr Bouchal \email{pbouchal@gmail.com} (\href{https://orcid.org/0000-0002-0471-716X}{ORCID})

}
\keyword{internal}
