% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kraje}
\alias{kraje}
\title{DATASET_TITLE}
\format{
A data frame with 14 rows and 4 variables:
\describe{
\item{\code{kraj_kod}}{character. DESCRIPTION.}
\item{\code{kraj_nazev}}{character. DESCRIPTION.}
\item{\code{kraj_kod_nuts3}}{character. DESCRIPTION.}
\item{\code{kraj_kod_ispv}}{character. DESCRIPTION.}
}
}
\usage{
kraje
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
