% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_load_cr.R
\name{pv_cr_monthlypay_education}
\alias{pv_cr_monthlypay_education}
\title{Load country-wide data on monthly earnings by education}
\usage{
pv_cr_monthlypay_education(path, sheet = 3)
}
\arguments{
\item{path}{path(s) to file(s), Will be file with "CR_{YYQ}D and either "PLS" or "MZS" in the name.}

\item{sheet}{sheet number; you should be able to leave this as default (3) if using files downloaded from ISPV}
}
\value{
a tibble
}
\description{
Loads data on monthly earnings by education level from one or more local Excel files downloaded from ISPV links retrieved by \code{pv_list_reg()}.
}
\examples{
pv_cr_monthlypay_education(system.file("extdata", "CR_204_PLS.xlsx", package = "ispv"))
pv_cr_monthlypay_education(system.file("extdata", "CR_204_MZS.xlsx", package = "ispv"))
}
