% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_load_cr.R
\name{pv_cr_monthlypay_isco}
\alias{pv_cr_monthlypay_isco}
\title{Load country-wide data by ISCO}
\usage{
pv_cr_monthlypay_isco(path, sheet = 1)
}
\arguments{
\item{path}{path to Excel input file provided by ISPV.}

\item{sheet}{which sheet to open. Will be 1 in files with only one sheet ("CR_204_MZS_M8r.xlsx")
and 8 in comprehensive file ("CR_204_MZS.xlsx")}
}
\value{
a tibble
}
\description{
Load country-wide data by ISCO
}
\examples{
pv_cr_monthlypay_isco(system.file("extdata", "CR_204_MZS_M8r.xlsx", package = "ispv"), 1)
}
