% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_load_cr.R
\name{pv_cr_monthlypay_nace}
\alias{pv_cr_monthlypay_nace}
\title{Load country-wide data by NACE}
\usage{
pv_cr_monthlypay_nace(path, sheet = 5)
}
\arguments{
\item{path}{path to file, filename should be in the form e.g. CR_YYQ_MZS.xlsx".}

\item{sheet}{Sheet number, defaults to 5}
}
\value{
a tibble
}
\description{
Load country-wide data by NACE (single-letter industry codes)
}
\examples{
pv_cr_monthlypay_nace(system.file("extdata", "CR_204_MZS.xlsx", package = "ispv"))
pv_cr_monthlypay_nace(system.file("extdata", "CR_204_PLS.xlsx", package = "ispv"))
}
