% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_list.R
\name{pv_list_reg}
\alias{pv_list_reg}
\title{Get links to Excel files with regional data on ISPV website for a given year}
\usage{
pv_list_reg(year = NULL, base_url = NULL, user_agent = NULL)
}
\arguments{
\item{year}{Year, defaults to scraping page for latest available data. String or numeric in YYYY format of length >= 1.}

\item{base_url}{base url, defaults to \url{https://ispv.cz}}

\item{user_agent}{User agent string, defaults to package URL on Github (https://github.com/petrbouchal/ispv).}
}
\value{
a tibble with file name, URL and year. File name can be used to determine the time period and region.
}
\description{
This function scrapes the relevant year's archive page for links to Excel files with
regional data and returns the result in a tibble.
}
\examples{
\dontrun{
pv_list_reg(year = 2019:2020)
}
}
