% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_load_reg.R
\name{pv_load_reg}
\alias{pv_load_reg}
\title{Load ISPV Excel files with regional data}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
pv_load_reg(path)
}
\arguments{
\item{path}{path to one or more Excel files (character vector length one or more).}
}
\value{
a tibble, see Format for details.
}
\description{
Loads one or more local Excel files downloaded from ISPV links retrieved by \code{pv_list_reg()}.
}
\examples{
pv_load_reg(system.file("extdata", "Kar_204_pls.xlsx", package = "ispv"))
}
