% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_load_reg.R
\name{pv_reg_monthlypay_education}
\alias{pv_reg_monthlypay_education}
\title{Load regional data on monthly earnings by education}
\usage{
pv_reg_monthlypay_education(path, sheet = 2)
}
\arguments{
\item{path}{path(s) to file(s), Will be file with "{Reg}_{YYQ}" and either "PLS" or "MZS" in the name.}

\item{sheet}{sheet number; you should be able to leave this as default (2) if using files downloaded from ISPV}
}
\value{
a tibble
}
\description{
Loads data on monthly earnings by ISCO code from one or more local Excel files downloaded from ISPV links retrieved by \code{pv_list_reg()}.
}
\examples{
pv_reg_monthlypay_education(system.file("extdata", "Kar_204_pls.xlsx", package = "ispv"))
pv_reg_monthlypay_education(system.file("extdata", "Kar_204_mzs.xlsx", package = "ispv"))
}
