% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kraje}
\alias{kraje}
\title{Codelist (registry) of regions}
\format{
A data frame with 14 rows and 4 variables:
\describe{
\item{\code{kraj_id_ispv}}{character. Internal region code, appears in Excel file name}
\item{\code{kraj_id}}{character. non-NUTS ID of kraj (region, NUTS3).}
\item{\code{kraj_name}}{character. Czech name of kraj (region, NUTS3).}
\item{\code{kraj_id_nuts3}}{character. NUTS ID of kraj (region, NUTS3)..}
}
}
\usage{
kraje
}
\description{
Used in \code{pv_load_kraje()} to add proper region IDs and names.
}
\keyword{datasets}
