% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_load_cr.R
\name{pv_cr_monthlypay_isco}
\alias{pv_cr_monthlypay_isco}
\title{Load country-wide data by ISCO}
\usage{
pv_cr_monthlypay_isco(path, sheet = 1)
}
\arguments{
\item{path}{path(s) to file(s), Will be file with "CR_{YYQ}D and either "PLS" or "MZS" in the name.}

\item{sheet}{which sheet to open. Will be 1 (the default) in files with only one sheet ("CR_204_MZS_M8r.xlsx")
and 7 in comprehensive file ("CR_204_MZS.xlsx")}
}
\value{
a tibble
}
\description{
Load country-wide data by ISCO  (profession classification) at 4 and 5 digits
}
\examples{
pv_cr_monthlypay_isco(system.file("extdata", "CR_204_MZS_M8r.xlsx", package = "ispv"), 1)
pv_cr_monthlypay_isco(system.file("extdata", "CR_204_MZS.xlsx", package = "ispv"), 7)
pv_cr_monthlypay_isco(system.file("extdata", "CR_204_PLS_M8r.xlsx", package = "ispv"), 1)
pv_cr_monthlypay_isco(system.file("extdata", "CR_204_PLS.xlsx", package = "ispv"), 7)
}
