% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_load_cr.R
\name{pv_cr_monthlypay_nace}
\alias{pv_cr_monthlypay_nace}
\title{Load country-wide data by NACE}
\usage{
pv_cr_monthlypay_nace(path, sheet = 5)
}
\arguments{
\item{path}{path(s) to file(s), Will be file with "CR_{YYQ}D and either "PLS" or "MZS" in the name.}

\item{sheet}{sheet number; you should be able to leave this as default (5) if using files downloaded from ISPV}
}
\value{
a tibble
}
\description{
Load country-wide data by NACE (single-letter industry codes)
}
\examples{
pv_cr_monthlypay_nace(system.file("extdata", "CR_204_MZS.xlsx", package = "ispv"))
pv_cr_monthlypay_nace(system.file("extdata", "CR_204_PLS.xlsx", package = "ispv"))
}
